recomendation-system/vector-creation-script/requirements.txt
# Worker node dependencies (runs on laptop)
sentence-transformers>=2.7.0
requests>=2.31.0
python-dotenv>=1.0.0
torch>=2.0.0
